requirements.txt
pytest==6.0.1
allure_python_commons==2.8.18
selenium==3.141.0
pynput==1.7.1
